#40 fixed conversion to ditamap
