+++ src/main/java/com/oxygenxml/resources/batch/converter/converters/WordToDITAConverter.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.oxygenxml.resources.batch.converter.UserInputsProvider;
+import com.oxygenxml.resources.batch.converter.persister.OptionTags;
 
 /**
  * Converter implementation for Word to DITA.
@@ -20,7 +21,11 @@ public class WordToDITAConverter extends PipelineConverter {
     List<Converter> converters = new ArrayList<Converter>();
     converters.add(new WordToXHTMLConverter());
     converters.add(new HtmlToDitaConverter());
-    converters.add(createDitaMapConverter(userInputsProvider));
+    Boolean shoultCreateDitaMap = userInputsProvider.getAdditionalOptionValue(
+        OptionTags.CREATE_DITA_MAP_FROM_WORD);
+    if(shoultCreateDitaMap != null && shoultCreateDitaMap) {
+      converters.add(new MapWithTopicsConverter());
+    }
 
     return converters.toArray(new Converter[0]);
   }
